Fixed initialization of gem
